Memcached test implementation complete
