+++ package.json
@@ -37,14 +37,16 @@
     "jsonwebtoken": "^8.5.1",
     "jwks-rsa": "^1.12.0",
     "jwt-decode": "^3.1.2",
-    "memcached-mock": "^0.1.0",
-    "node-fetch": "^2.6.1"
+    "memcached-node": "^0.1.0",
+    "node-fetch": "^2.6.1",
+    "typemoq": "^2.1.0"
   },
   "devDependencies": {
     "@types/jest": "^26.0.15",
     "@types/node": "^12.12.67",
     "baretest": "^2.0.0",
     "jest": "^26.6.3",
+    "memcached-mock": "^0.1.0",
     "ts-jest": "^26.4.4",
     "typescript": "^4.0.3"
   }
