+++ src/endpoints/cache.ts
@@ -1,9 +1,10 @@
 import {FastifyInstance} from "fastify";
-import {Auth0JwtVerifier, removeUserFromCache} from "../auth0";
+import {Auth0JwtVerifier} from "../auth0";
 import {AuthenticatedRequest} from "./endpointUtils";
 import {emptyCacheSchema} from "./docs/cacheSchemas";
+import {Client} from "memjs";
 
-export function createCacheEndpoint(app: FastifyInstance, verifyJwt: Auth0JwtVerifier) {
+export function createCacheEndpoint(app: FastifyInstance, verifyJwt: Auth0JwtVerifier, cache: Client) {
   app.post<AuthenticatedRequest>(
     "/cache/empty",
     {schema: emptyCacheSchema},
@@ -11,8 +12,13 @@ export function createCacheEndpoint(app: FastifyInstance, verifyJwt: Auth0JwtVer
       let {userAppId} = await verifyJwt(request);
       if (!userAppId) {
         reply.unauthorized("Must be logged in to use this endpoint");
+        return;
       } else {
-        removeUserFromCache(request)
+        await cache.flush();
+        return {
+          status: "Cache emptied",
+          date: Date.now()
+        }
       }
     }
   );
