+++ src/index.ts
@@ -1,5 +1,7 @@
 import fastify, {FastifyInstance} from 'fastify';
 import fastifySensible from "fastify-sensible";
+import MemcachePlus from "memcache-plus";
+
 import createPingEndpoint from './endpoints/ping';
 import { addRoutes } from './utils';
 import { createBusinessesEndpoint} from "./endpoints/businesses";
@@ -8,10 +10,11 @@ import {ProductionDataLayer} from "./database/productionDataLayer";
 import {createFiltersEndpoint} from "./endpoints/filters";
 import {registerCorsHandler} from "./cors";
 import {registerSwagger} from "./swagger";
-import {verifyJwt} from "./auth0";
+import {getJwtVerifier} from "./auth0";
 import {createEditEndpoint} from "./endpoints/editRequest";
 import {productionFirestore} from "./database/firestore";
 import createUsersEndpoint from "./endpoints/users";
+import {createCacheEndpoint} from "./endpoints/cache";
 
 let productionDataLayer = new ProductionDataLayer(productionFirestore)
 const port = Number(process.env.PORT || 8080);
@@ -20,14 +23,17 @@ server.register(fastifySensible);
 registerSwagger(server);
 registerCorsHandler(server);
 
+let productionCache = new MemcachePlus();
+let prodJwtVerifier = getJwtVerifier(productionCache);
 addRoutes(
   server,
   createPingEndpoint,
-  (app: FastifyInstance) => createFiltersEndpoint(app, productionDataLayer, verifyJwt),
-  (app: FastifyInstance) => createRegionsEndpoint(app, productionDataLayer, verifyJwt),
-  (app: FastifyInstance) => createBusinessesEndpoint(app, productionDataLayer, verifyJwt),
-  (app: FastifyInstance) => createEditEndpoint(app, productionDataLayer, verifyJwt),
-  (app: FastifyInstance) => createUsersEndpoint(app, verifyJwt)
+  (app: FastifyInstance) => createFiltersEndpoint(app, productionDataLayer, prodJwtVerifier),
+  (app: FastifyInstance) => createRegionsEndpoint(app, productionDataLayer, prodJwtVerifier),
+  (app: FastifyInstance) => createBusinessesEndpoint(app, productionDataLayer, prodJwtVerifier),
+  (app: FastifyInstance) => createEditEndpoint(app, productionDataLayer, prodJwtVerifier),
+  (app: FastifyInstance) => createUsersEndpoint(app, prodJwtVerifier),
+  (app: FastifyInstance) => createCacheEndpoint(app, prodJwtVerifier, productionCache)
 );
 
 server.listen(port, '::', (err, address) => {
